For #10734,fix codestyle , move extractor and extractorManager to core directory
