+++ api/src/main/java/com/alibaba/nacos/api/config/remote/request/AbstractConfigRequest.java
@@ -26,37 +26,37 @@ import com.alibaba.nacos.api.remote.request.Request;
  * @version $Id: ConfigCommonRequest.java, v 0.1 2020年07月13日 9:05 PM liuzunfei Exp $
  */
 public abstract class AbstractConfigRequest extends Request {
-
+    
     private String dataId;
-
+    
     private String group;
-
+    
     private String tenant;
-
+    
     public String getDataId() {
         return dataId;
     }
-
+    
     public void setDataId(String dataId) {
         this.dataId = dataId;
     }
-
+    
     public String getGroup() {
         return group;
     }
-
+    
     public void setGroup(String group) {
         this.group = group;
     }
-
+    
     public String getTenant() {
         return tenant;
     }
-
+    
     public void setTenant(String tenant) {
         this.tenant = tenant;
     }
-
+    
     @Override
     public String getModule() {
         return Constants.Config.CONFIG_MODULE;
