+++ api/src/main/java/com/alibaba/nacos/api/config/remote/request/ConfigRemoveRequest.java
@@ -23,20 +23,20 @@ package com.alibaba.nacos.api.config.remote.request;
  * @version $Id: ConfigRemoveRequest.java, v 0.1 2020年07月16日 4:31 PM liuzunfei Exp $
  */
 public class ConfigRemoveRequest extends AbstractConfigRequest {
-
+    
     String tag;
-
+    
     public ConfigRemoveRequest() {
-
+    
     }
-
+    
     public ConfigRemoveRequest(String dataId, String group, String tenant, String tag) {
         super.setDataId(dataId);
         super.setGroup(group);
         super.setTenant(tenant);
         this.tag = tag;
     }
-
+    
     /**
      * Getter method for property <tt>tag</tt>.
      *
@@ -45,7 +45,7 @@ public class ConfigRemoveRequest extends AbstractConfigRequest {
     public String getTag() {
         return tag;
     }
-
+    
     /**
      * Setter method for property <tt>tag</tt>.
      *
@@ -54,5 +54,5 @@ public class ConfigRemoveRequest extends AbstractConfigRequest {
     public void setTag(String tag) {
         this.tag = tag;
     }
-
+    
 }
