+++ common/src/main/java/com/alibaba/nacos/common/paramcheck/ParamInfo.java
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package com.alibaba.nacos.common.paramcheck.utils;
+package com.alibaba.nacos.common.paramcheck;
 
 import java.util.Map;
 
@@ -25,91 +25,91 @@ import java.util.Map;
  */
 public class ParamInfo {
     private String namespaceShowName;
-
+    
     private String namespaceId;
-
+    
     private String dataId;
-
+    
     private String serviceName;
-
+    
     private String group;
-
+    
     private String cluster;
-
+    
     private String ip;
-
+    
     private String port;
-
+    
     private Map<String, String> metadata;
-
+    
     public String getNamespaceShowName() {
         return namespaceShowName;
     }
-
+    
     public void setNamespaceShowName(String namespaceShowName) {
         this.namespaceShowName = namespaceShowName;
     }
-
+    
     public String getNamespaceId() {
         return namespaceId;
     }
-
+    
     public void setNamespaceId(String namespaceId) {
         this.namespaceId = namespaceId;
     }
-
+    
     public String getDataId() {
         return dataId;
     }
-
+    
     public void setDataId(String dataId) {
         this.dataId = dataId;
     }
-
+    
     public String getServiceName() {
         return serviceName;
     }
-
+    
     public void setServiceName(String serviceName) {
         this.serviceName = serviceName;
     }
-
+    
     public String getGroup() {
         return group;
     }
-
+    
     public void setGroup(String group) {
         this.group = group;
     }
-
+    
     public String getCluster() {
         return cluster;
     }
-
+    
     public void setCluster(String cluster) {
         this.cluster = cluster;
     }
-
+    
     public String getIp() {
         return ip;
     }
-
+    
     public void setIp(String ip) {
         this.ip = ip;
     }
-
+    
     public String getPort() {
         return port;
     }
-
+    
     public void setPort(String port) {
         this.port = port;
     }
-
+    
     public Map<String, String> getMetadata() {
         return metadata;
     }
-
+    
     public void setMetadata(Map<String, String> metadata) {
         this.metadata = metadata;
     }
